% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_object.R
\name{add_object}
\alias{add_object}
\title{Function to add an object to the JSON}
\usage{
add_object(json_obj, input_list)
}
\arguments{
\item{input_list}{A list of fields corresponding to entries. Must include
\code{name,description,Units}}
}
\description{
Function to add an object to the JSON
}
