% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{CaribbeanFisheryManagementCouncil}
\alias{CaribbeanFisheryManagementCouncil}
\title{Caribbean Fishery Management Council (CFMC)}
\format{
\describe{
}
}
\source{
Magnuson-Stevens Act, sec. 302 (16 USC 1852)
}
\usage{
CaribbeanFisheryManagementCouncil
}
\description{
One of eight US Regional Fishery Management Councils established by the
Magnuson-Stevens Act. Each Council is responsible for preparing and submitting
to the Secretary of Commerce proposed fishery management plans and amendments
to those plans, as well as proposed regulations that the Council deems
necessary and appropriate to implement the plan, for fisheries within the
Council's specified geographic region that require conservation and management.
The Caribbean Council consists of the U.S. Virgin Islands and the Commonwealth
of Puerto Rico and generally has authority over the fisheries in the Caribbean
Sea and Atlantic Ocean seaward of the Virgin Islands and Puerto Rico.
}
\keyword{datasets}
