% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catch.R
\docType{data}
\name{Catch}
\alias{Catch}
\title{Catch}
\format{
\describe{
\item{Examples}{recreational catch}
\item{Rationale}{Landings and catch are sometimes thought to be
interchangeable but they are not given that catch can also include
bycatch or unwanted catch.}
\item{Alternatives}{total mortality, harvest, total removals}
\item{Range of possible values}{0--Inf}
\item{Units}{mt, numbers}
}
}
\usage{
Catch
}
\description{
Catch is the total quantity of fish, measured in weight or numbers of fish, taken
in commercial, recreational, subsistence, tribal, and other fisheries. Catch includes
fish that are retained for any purpose, as well as mortality of fish that are discarded.
}
\keyword{datasets}
