% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchPerUnitEffort.R
\docType{data}
\name{CatchPerUnitEffort}
\alias{CatchPerUnitEffort}
\title{Catch Per Unit Effort (CPUE)}
\format{
\describe{
\item{Examples}{fishery-independent CPUE}
\item{Rationale}{Many models use the term index for a survey and CPUE for
fishery-dependent data but they are typically entered into an input
file in the same way and the naming should be standardized to a single
overarching term. The overarching term would facilitate the inclusion of
environmental data thought to index at least a portion of the population
dynamics as well as fishery data. Environmental data does not necessarily
have a catch or unit of effort associated with it, and thus, the term CPUE
does not apply. Often CPUE data are thought to be a relative index of
abundance and would never be considered absolute such as can be the case
for an acoustic survey that does not kill the population but rather creates
a metric of its abundance potentially over the entire area in which it is
thought to inhabit. Where, an absolute index of abundance would be assumed
to have a catchability of one.}
\item{Alternatives}{catch rate, index of abundance, standardized fishery
catch time series, catch per effort, fishing success, availability.}
\item{Range of possible values}{0--Inf}
\item{Units}{}
}
}
\usage{
CatchPerUnitEffort
}
\description{
The quantity of fish caught (in number or in weight)
with one standard unit of fishing effort; e.g. number of
fish taken per 1,000 hooks per day or weight of fish,
in tons, taken per hour of trawling. CPUE is often
considered an index of fish biomass (or abundance).
Sometimes referred to as catch rate. CPUE may be used
as a measure of economic efficiency of fishing as well as
an index of fish abundance.
}
\keyword{datasets}
