% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_create_contributing_file.R
\name{create_acronym_roxygen}
\alias{create_acronym_roxygen}
\title{Use Roxygen formatting for a new entry into {FishDictionary}}
\usage{
create_acronym_roxygen(acronym, meaning, definition, source)
}
\arguments{
\item{acronym}{Acronym}

\item{meaning}{Meaning of acronym}

\item{definition}{Definition of acronym}

\item{source}{Source of definition}
}
\value{
A string of Roxygen formatted text for the new entry
}
\description{
Use Roxygen formatting for a new entry into {FishDictionary}
}
\examples{
{
create_acronym_roxygen(
  acronym = "CH",
  meaning = "Critical Habitat",
  definition = "The specific areas within the geographical area occupied by the species.",
  source = "Endangered Species Act sec. 3(5) (16 USC 1532(5))"
)
}
}
