% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_create_contributing_file.R
\name{create_contrib_file}
\alias{create_contrib_file}
\title{Contribute to fish dictionary using csv sheet of new entries}
\usage{
create_contrib_file(file, dir = "./R", FileName = "asarAcronyms.R")
}
\arguments{
\item{file}{csv file containing the acronym (column name = "Acronym"),
meaning (column name = "Meaning"), definition (column name = "Definition"),
and source (column name = "Source") of each term}

\item{dir}{file path where the new file is created. Default to the R folder
in the FishDictionary repository when working in a project}

\item{FileName}{A string describing the name of the output file created.}
}
\value{
A .R file with Roxygen formatted entries for each term in the input csv file.
}
\description{
Contribute to fish dictionary using csv sheet of new entries
}
