% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{DataQualityAct}
\alias{DataQualityAct}
\title{Data Quality Act (DQA)}
\format{
\describe{
}
}
\source{
NOAA Fisheries, Policy on the Data Quality Act, PD 04-108 (June 27, 2012; renewed July 2014); https://media.fisheries.noaa.gov/dam-migration/04-108.pdf
}
\usage{
DataQualityAct
}
\description{
The Data Quality Act (also referred to as the Information Quality Act),
requires federal agencies like NOAA Fisheries to ensure the quality,
objectivity, utility, and integrity of the information they disseminate.
}
\keyword{datasets}
