% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{EcosystemBasedFisheriesManagement}
\alias{EcosystemBasedFisheriesManagement}
\title{Ecosystem-Based Fisheries Management (EBFM)}
\format{
\describe{
}
}
\source{
NOAA Fisheries; https://www.fisheries.noaa.gov/insight/understanding-ecosystem-based-fisheries-management
}
\usage{
EcosystemBasedFisheriesManagement
}
\description{
A systematic approach to fisheries management in a geographically specified
area that contributes to: the resilience and sustainability of the ecosystem;
recognizes the physical, biological, economic, and social interactions among
the affected fishery-related components of the ecosystem, including humans; and
seeks to optimize benefits among a diverse set of societal goals.
}
\keyword{datasets}
