% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{EnvironmentalImpactStatement}
\alias{EnvironmentalImpactStatement}
\title{Environmental Impact Statement (EIS)}
\format{
\describe{
}
}
\source{
42 U.S.C. 4332(2)(C) and  42 US.C. 4336(b)(1); NOAA NEPA Procedures at 12 (June 30, 2025); NMFS, Revised MSA-NEPA Procedures, PD 09-103 at 5-6 (revised June 30, 2025); CEQ, A Citizen’s Guide to NEPA at 12 (Jan. 2021); NMFS, Operational Guidelines, PD 01-101-03, App. 2 at 25 (revised Oct. 25, 2017)
}
\usage{
EnvironmentalImpactStatement
}
\description{
A detailed statement required by NEPA for any major Federal action
significantly affecting the quality of the human environment, which addresses
the reasonably foreseeable environmental effects of a proposed action and
reasonable alternatives to that action.
}
\keyword{datasets}
