% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishingMortalityAtMaximumSustainableYield.R
\docType{data}
\name{FishingMortalityAtMaximumSustainableYield}
\alias{FishingMortalityAtMaximumSustainableYield}
\title{Fishing mortality at maximum sustainable yield (FMSY)
A biological reference point for fisheries management
based on the fishing mortality rate (F) that achieves the maximum
sustainable yield (MSY).}
\format{
A schema with the following fields:
\describe{
\item{Examples}{}
\item{Rationale}{The notation used for reference points of commonly
calculated values is inconsistent. For example, some use subscripts
where the subscript can be in either all capital letters or all
lower-case letters, some use lower-case letters for the type of reference
point, and some use all capital letters.}
\item{Alternatives: Fproxy, F_msy, Fmsy}
\item{Range of possible values: 0--Inf}
\item{Units}{\code{time^{-1}}}
}
}
\usage{
FishingMortalityAtMaximumSustainableYield
}
\description{
Fishing mortality at maximum sustainable yield (FMSY)
A biological reference point for fisheries management
based on the fishing mortality rate (F) that achieves the maximum
sustainable yield (MSY).
}
\keyword{datasets}
