% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\docType{data}
\name{Forecast}
\alias{Forecast}
\title{Forecast}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
Forecast
}
\description{
A probabilistic statement about future events.
}
\details{
@format A schema with the following fields:
\describe{
\item{Examples}{short-term forecast}
\item{Rationale}{Future weather is a forecast because it does not require
an if statement, whereas future spawning biomass is a projection because
it depends on an assumption about future catches.}
\item{Alternatives}{projection, prediction}
\item{Range of possible values}{}
\item{Units}{}
}
}
\keyword{datasets}
