% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{GreaterAtlanticRegionalFisheriesOffice}
\alias{GreaterAtlanticRegionalFisheriesOffice}
\title{Greater Atlantic Regional Fisheries Office (GARFO)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/greater-atlantic-regional-fisheries-office
}
\usage{
GreaterAtlanticRegionalFisheriesOffice
}
\description{
One of NOAA Fisheries' five Regional Offices that supports conservation and
management of fisheries and living marine resources. The Greater Atlantic
Region covers approximately 100,000 square miles of the Northwest Atlantic.
This vast area includes the temperate, complex large marine ecosystems from
Maine to Cape Hatteras, North Carolina, as well as the Great Lakes and rivers
and estuaries within this range.
}
\keyword{datasets}
