% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{MagnusonStevensFisheryConservationandManagementReauthorizationActof2006}
\alias{MagnusonStevensFisheryConservationandManagementReauthorizationActof2006}
\title{Magnuson-Stevens Fishery Conservation and Management Reauthorization Act of 2006 (MSRA)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/topic/laws-policies/magnuson-stevens-act
}
\usage{
MagnusonStevensFisheryConservationandManagementReauthorizationActof2006
}
\description{
An act, signed into law in January 2007, that made significant revisions to the
MSA to further refine and strengthen fisheries science, management, and
conservation.
}
