% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hex.R
\name{make_hex}
\alias{make_hex}
\title{Create a hexagonal sticker for the 'fishdictionary' package.}
\usage{
make_hex()
}
\value{
The function does not return any value but creates a hexagonal sticker image.
}
\description{
Create a hexagonal sticker for the 'fishdictionary' package.
}
