% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mass.R
\name{Mass}
\alias{Mass}
\title{Mass}
\format{
\describe{
\item{Examples}{average mass-at-age}
\item{Rationale}{The unit of measurement for the mass of individual
fish can vary by country or even between sampling organizations. For
example, some organizations measure fish mass in pounds rather than
in kilograms and sometimes kilograms are reported using grams. The metric
system is an internationally agreed-upon system of measurement and, more
specifically, the International System of Units (SI) is used in science
and should be used for fisheries science. Thus, measurements of mass
should be reported in kilograms (kg).}
\item{Alternatives}{weight}
\item{Range of possible values}{0--Inf}
\item{Units}{kg}
}
}
\usage{
Mass()
}
\description{
The individual body mass of a fish that may come from direct
measurements or an average across a sample.
}
