% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumSustainableYield.R
\docType{data}
\name{MaximumSustainableYield}
\alias{MaximumSustainableYield}
\title{Maximum Sustainable Yield}
\format{
\describe{
\item{Examples}{}
\item{Rationale}{Many management frameworks are based on calculations
of MSY despite difficulty in its estimation. Metrics such as fishing rate
at MSY should be abbreviated using all upper case letters but this is not
common practice as of yet. For example, sometimes you will see \code{F_msy} or
some derivative thereof.}
\item{Alternatives}{msy, MSYproxy}
\item{Range of possible values}{0--Inf}
\item{Units}{mt, numbers}
}
}
\source{
insert URL
}
\usage{
MaximumSustainableYield
}
\description{
The largest long-term average catch or yield that can be taken from a stock or
stock complex under prevailing ecological, environmental conditions, and
fishery technological characteristics (e.g., gear selectivity) and the distribution
of catch among fleets. It can be defined using biological parameters
(e.g., natural mortality and stock productivity) or proxies (e.g., the maximum
yield at 40\% of unfished biomass). Units depend on the units used for the catch.
}
\keyword{datasets}
