% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{MigratoryBirdTreatyAct}
\alias{MigratoryBirdTreatyAct}
\title{Migratory Bird Treaty Act (MBTA)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/topic/laws-policies/more-laws
}
\usage{
MigratoryBirdTreatyAct
}
\description{
An Act that makes it illegal to take, possess, import, export, transport, sell,
purchase, barter, or offer for sale, purchase, or barter, any migratory bird,
or the parts (feathers), nests, or eggs of such a bird except under the terms
of a valid federal permit.
}
\keyword{datasets}
