% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaturalLogOfUnfishedRecruitment.R
\docType{data}
\name{NaturalLogOfUnfishedRecruitment}
\alias{NaturalLogOfUnfishedRecruitment}
\title{Natural log of unfished recruitment}
\format{
\describe{
\item{Examples}{}
\item{Rationale}{The notation for this value is vast, which can be seen
in the number of alternatives. It is most important to remember that the
units of the measure of unfished recruitment before the natural log is taken
is in 1000’s of recruits or fish. Both ln and log are used even though the
transformation is for sure not a base-10 transformation and is rather a
transformation on the natural-log scale.}
\item{Synonym or Similar Terms}{logR0, \code{logR_0}, log(R0), R0}
\item{Range of possible values}{0--Inf}
\item{Units}{ln(numbers x 1000)}
}
}
\source{
cite here
}
\usage{
NaturalLogOfUnfishedRecruitment
}
\description{
Also known as lnR0. The average number of fish recruited to the population
under equilibrium conditions on the natural-log scale.
}
\keyword{datasets}
