% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{NorthPacificFisheryManagementCouncil}
\alias{NorthPacificFisheryManagementCouncil}
\title{North Pacific Fishery Management Council (NPFMC)}
\format{
\describe{
}
}
\source{
Magnuson-Stevens Act, sec. 302 (16 USC 1852)
}
\usage{
NorthPacificFisheryManagementCouncil
}
\description{
One of eight US Regional Fishery Management Councils established by the
Magnuson-Stevens Act. Each Council is responsible for preparing and submitting
to the Secretary of Commerce proposed fishery management plans and amendments
to those plans, as well as proposed regulations that the Council deems
necessary and appropriate to implement the plan, for fisheries within the
Council's specified geographic region that require conservation and management.
The North Pacific Council consists of the States of Alaska, Washington, and
Oregon and generally has authority over the fisheries in the Arctic Ocean,
Bering Sea, and Pacific Ocean seaward of Alaska.
}
\keyword{datasets}
