% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumbersAtAge.R
\name{NumbersAtAge}
\alias{NumbersAtAge}
\title{Numbers at age}
\format{
\describe{
\item{Examples}{}
\item{Rationale}{}
\item{Alternatives}{}
\item{Range of possible values}{0--Inf}
\item{Units}{numbers of fish}
}
}
\usage{
NumbersAtAge()
}
\description{
The estimated abundance of fish (in numbers) at each population age
}
