% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{PacificFisheryManagementCouncil}
\alias{PacificFisheryManagementCouncil}
\title{Pacific Fishery Management Council (PFMC)}
\format{
\describe{
}
}
\source{
Magnuson-Stevens Act, sec. 302 (16 USC 1852)
}
\usage{
PacificFisheryManagementCouncil
}
\description{
One of eight US Regional Fishery Management Councils established by the
Magnuson-Stevens Act. Each Council is responsible for preparing and submitting
to the Secretary of Commerce proposed fishery management plans and amendments
to those plans, as well as proposed regulations that the Council deems
necessary and appropriate to implement the plan, for fisheries within the
Council's specified geographic region that require conservation and management.
The Pacific Council consists of the States of California, Oregon, Washington,
and Idaho and generally has authority over the fisheries in the Pacific Ocean
seaward of these States.
}
\keyword{datasets}
