% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{PacificIslandsRegionalOffice}
\alias{PacificIslandsRegionalOffice}
\title{Pacific Islands Regional Office (PIRO)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/pacific-islands-regional-office
}
\usage{
PacificIslandsRegionalOffice
}
\description{
One of NOAA Fisheries' five Regional Offices that supports conservation and
management of fisheries and living marine resources. The Pacific Islands Region
encompasses more than 1.7 million square nautical miles of ocean. The specific
land areas in this jurisdiction are Hawaiʻi, American Samoa, Guam, the
Commonwealth of the Northern Mariana Islands, and the Pacific Remote Island
Areas of Kingman Reef; Howland, Baker, Jarvis, and Wake islands; and Johnston
and Palmyra atolls.
}
\keyword{datasets}
