% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlusGroup.R
\docType{data}
\name{PlusGroup}
\alias{PlusGroup}
\title{Plus group}
\format{
A schema with the following fields:
\describe{
\item{Examples}{}
\item{Rationale}{}
\item{Alternatives}{}
\item{Range of possible values}{0--Inf}
\item{Units}{whole year age}
}
}
\usage{
PlusGroup
}
\description{
fish of a designated age and older where the life history of all ages
in the group are treated similarly because the ability to predict age
from length is difficult given the lack of increase in length or there is
limited data to define differences.
}
\keyword{datasets}
