% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection.R
\docType{data}
\name{Projection}
\alias{Projection}
\title{Projection}
\format{
\describe{
\item{Examples}{short-term projection}
\item{Rationale}{Future weather is a forecast because it does not require
an if statement, whereas future spawning biomass is a projection because
it depends on an assumption about future catches.}
\item{Synonym or Similar Terms}{forecast, prediction}
\item{Range of possible values}{}
\item{Units}{}
}
}
\usage{
Projection
}
\description{
A probabilistic statement about future events based on what might happen.
}
\keyword{datasets}
