% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{read_in}
\alias{read_in}
\title{Function to read in a txt and make it a data frame}
\usage{
read_in(filename = "inst/extdata/top20.txt")
}
\arguments{
\item{filename}{name of the txt file to convert to JSON}
}
\description{
Function to read in a txt and make it a data frame
}
