% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruits.R
\docType{data}
\name{Recruits}
\alias{Recruits}
\title{Recruits}
\format{
\describe{
\item{Examples}{thousands of recruits, stock-recruitment relationship}
\item{Rationale}{Recruitment can also be used to describe fish reaching a
legal size and becoming vulnerable to harvesting, though we will label it
as such rather than recruitment to the stock. Age-0 fish are typically
referred to as recruits but documentation should state the age at which
fish are recruited because there is no standard.}
\item{Alternatives}{age-0 fish}
\item{Range of possible values}{0--Inf}
\item{Units}{numbers}
}
}
\source{
Lorenzen and Camp, 2019
}
\usage{
Recruits
}
\description{
The number of fish transitioning to a different life stage
(Lorenzen and Camp, 2019), typically from that of a pre-juvenile stage
that could experience density-dependent life-history processes to a
juvenile stage without density-dependent processes.
}
\keyword{datasets}
