% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{Regionalfisherymanagementorganization}
\alias{Regionalfisherymanagementorganization}
\title{Regional fishery management organization (RFMO)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/international-affairs/international-and-regional-fisheries-management-organizations
}
\usage{
Regionalfisherymanagementorganization
}
\description{
Treaty-based body whose objective is to ensure the sustainable conservation and
management of shared fish stocks and other living marine resources through
international cooperation.
}
\keyword{datasets}
