% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{RegulatoryFlexibilityAct}
\alias{RegulatoryFlexibilityAct}
\title{Regulatory Flexibility Act (RFA)}
\format{
\describe{
}
}
\source{
CRS, In Focus, The Regulatory Flexibility Act: An Overview (Aug. 16, 2021), https://www.congress.gov/crs-product/IF11900; SBA, Office of Advocacy, How to Comply with the RFA, https://advocacy.sba.gov/resources/regulatory-flexibility-act/rfa-basics/how-to-comply-with-the-regulatory-flexibility-act/; SBA, Office of Advocacy, How to Comply with the RFA (Aug. 2017), https://advocacy.sba.gov/wp-content/uploads/2019/06/How-to-Comply-with-the-RFA.pdf
}
\usage{
RegulatoryFlexibilityAct
}
\description{
An act that requires federal agencies to consider the effects of their
regulations on small businesses and other small entities.
}
\keyword{datasets}
