% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sex.R
\docType{data}
\name{Sex}
\alias{Sex}
\title{Sex}
\format{
\describe{
\item{Examples}{}
\item{Rationale}{Fisheries science is concerned about the reproductive
capabilities of fish, and thus, sex is more appropriate than gender because
sex is based on reproductive organs and biological functions that are
realized because of chromosomal composition (Torgrimson and Minson, 2005).}
\item{Alternatives}{gender}
\item{Range of possible values}{01, FMU, Female, Male, Transitioning,
Undetermined}
\item{Units}{}
}
}
\source{
Torgrimson and Minson, 2005
}
\usage{
Sex
}
\description{
The biological properties of an organism that help to distinguish it as a
male, female, or hermaphrodite.
}
\keyword{datasets}
