% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dd.r
\name{shiny_dd}
\alias{shiny_dd}
\title{Run the data dictionary app}
\usage{
shiny_dd(browse = TRUE)
}
\arguments{
\item{browse}{Logical. Use browser for running Shiny app.}
}
\description{
Run the data dictionary app
}
\examples{
\dontrun{
if (require(shiny)) {
  shiny_dd()
}
}
}
