% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{SouthAtlanticRegionalOffice}
\alias{SouthAtlanticRegionalOffice}
\title{South Atlantic Regional Office (SERO)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/southeast-regional-office
}
\usage{
SouthAtlanticRegionalOffice
}
\description{
One of NOAA Fisheries' five Regional Offices that supports conservation and
management of fisheries and living marine resources. The Southeast Region
covers nearly 20,000 miles of tidal coastline throughout the South Atlantic,
Gulf of America (formerly Gulf of Mexico), and Caribbean. This includes the
eight coastal states of North Carolina, South Carolina, Georgia, Florida,
Alabama, Mississippi, Louisiana, and Texas; the inland watershed states of
Arkansas, Iowa, Kansas, Kentucky, Missouri, Nebraska, New Mexico, Oklahoma, and
Tennessee; and the Commonwealth of Puerto Rico and the U.S. Virgin Islands.
}
\keyword{datasets}
