% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\docType{data}
\name{SouthwestFisheriesScienceCenter}
\alias{SouthwestFisheriesScienceCenter}
\title{Southwest Fisheries Science Center (SWFSC)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/southwest-fisheries-science-center
}
\usage{
SouthwestFisheriesScienceCenter
}
\description{
One of NOAA Fisheries' six Science Centers. The Southwest center provides
science to support the conservation and management of fisheries and living
marine resources across the California Current, throughout the Pacific Ocean,
and in the Southern Ocean off Antarctica.
}
\keyword{datasets}
