% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpawnerPerRecruit.R
\docType{data}
\name{SpawnerPerRecruit}
\alias{SpawnerPerRecruit}
\title{Spawner per recruit (SPR)}
\format{
\describe{
\item{Examples}{SPR30\% }
\item{Rationale}{Spawner per recruit and spawning potential ratio have the
same acronym of SPR.}
\item{Synonym or Similar Terms}{30\%SPR}
\item{Range of possible values}{0--1}
\item{Units}{}
}
}
\usage{
SpawnerPerRecruit
}
\description{
The ratio of the amount of spawning biomass produced by a cohort over its
lifespan under a specific fishing regime relative to the spawning biomass
produced over the cohort's lifespan under no fishing. The numerator will
always be smaller than the denominator because of mortality from fishing.
Only information about the life history and fishing is needed for the
calculation, whereas other reference points may need information about
recruitment, making it a common proxy for reference points based on maximum
sustainable yield.
}
\keyword{datasets}
