% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_Rd.R
\name{validate_Rd}
\alias{validate_Rd}
\title{Function to validate terms in the dictionary.}
\usage{
validate_Rd(Rd_file)
}
\arguments{
\item{Rd_file}{a Rd filename to use as input.}
}
\value{
a message if the term passes all validations.
}
\description{
The dictionary terms are written in R documentation (Rd) format.
This function reads the Rd file using the \code{tools::parse_Rd} function
validate the information in each section (e.g., examples, rationale,
range of possible values, etc.).
}
