% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{AlaskaFisheriesScienceCenter}
\alias{AlaskaFisheriesScienceCenter}
\title{Alaska Fisheries Science Center (AFSC)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center
}
\usage{
AlaskaFisheriesScienceCenter
}
\description{
One of NOAA Fisheries' six Science Centers. The Alaska science center provides
science to support the conservation and management of fisheries and living
marine resources across nearly 1.5 million square miles of federal waters
surrounding the state of Alaska.
}
