% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{AlaskaRegionalOffice}
\alias{AlaskaRegionalOffice}
\title{Alaska Regional Office (AKRO)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/alaska-regional-office
}
\usage{
AlaskaRegionalOffice
}
\description{
One of NOAA Fisheries' five Regional Offices that supports conservation and
management of fisheries and living marine resources. The Alaska Region includes
nearly 1.5 million square miles (70 percent of the total U.S. continental
shelf) of the North Pacific Ocean, including the Arctic Oceans (Chukchi and
Beaufort seas), Gulf of Alaska, and the eastern Bering Sea.
}
