% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catch.R
\docType{data}
\name{Catch}
\alias{Catch}
\title{Catch}
\format{
\describe{
\item{Examples}{recreational catch}
\item{Rationale}{Landings and catch are sometimes thought to be
interchangeable but they are not given that catch can also include
bycatch or unwanted catch.}
\item{Alternatives}{total mortality}
\item{Range of possible values}{0--Inf}
\item{Units}{mt, numbers}
}
}
\usage{
Catch
}
\description{
Everything that died due to fishing, i.e., both landed and discarded fish.
}
\keyword{datasets}
