% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\docType{data}
\name{Cohort}
\alias{Cohort}
\title{Cohort}
\format{
A schema with the following fields:
\describe{
\item{Examples}{}
\item{Rationale}{}
\item{Synonym or Similar Terms}{recruit}
\item{Range of possible values}{}
\item{Units}{}
}
}
\source{
see
}
\usage{
Cohort
}
\description{
Individuals of a stock born in the same spawning season.
}
\keyword{datasets}
