% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalHabitat.R
\name{CriticalHabitat}
\alias{CriticalHabitat}
\title{Critical Habitat (CH)}
\format{
\describe{
}
}
\source{
Endangered Species Act sec. 3(5) (16 USC 1532(5))
}
\usage{
CriticalHabitat
}
\description{
When applied for a threatened or endangered species: the specific areas within the geographical area occupied by the species, at the time it is listed, on which are found those physical or biological features essential to the conservation of the species and which may require special management considerations or protection; and specific areas outside the geographical area occupied by the species at the time it is listed, upon a determination by the Secretary that such areas are essential for the conservation of the species.
}
