% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{EssentialFishHabitat}
\alias{EssentialFishHabitat}
\title{Essential Fish Habitat (EFH)}
\format{
\describe{
}
}
\source{
Magnuson-Stevens Act, sec. 3(10) (16 USC 1802(10))
}
\usage{
EssentialFishHabitat
}
\description{
Those waters and substrate necessary to fish for spawning, breeding, feeding or
growth to maturity.
}
