% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{EssentialFishHabitatHabitatAreaofParticularConcern}
\alias{EssentialFishHabitatHabitatAreaofParticularConcern}
\title{Essential Fish Habitat- Habitat Area of Particular Concern (EFH-HAPC)}
\format{
\describe{
}
}
\source{
NOAA Fisheries; 50 CFR 600.815(a)(8)
}
\usage{
EssentialFishHabitatHabitatAreaofParticularConcern
}
\description{
A subset of Essential Fish Habitat (EFH) that is designated as high priority
for conservation and management based on: the importance of the ecological
functions it provides; its sensitivity to human-induced environmental
degradation; whether, and to what extent, development activities are stressing
the habitat; and the rarity of the habitat type.
}
