% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{FisheryEcosystemPlan}
\alias{FisheryEcosystemPlan}
\title{Fishery Ecosystem Plan (FEP)}
\format{
\describe{
}
}
\source{
NMFS, Ecosystem-Based Fisheries Management Policy, PD 01-120 (May 23, 2016) (https://www.fisheries.noaa.gov/s3//2024-02/Revised-EBFM-Policy-FINAL-2.12.24-508-signed-JC.pdf); NOAA Fisheries (https://www.fisheries.noaa.gov/topic/ecosystems/fishery-ecosystem-plans)
}
\usage{
FisheryEcosystemPlan
}
\description{
A strategic planning document that describes and integrates ecosystem goals,
objectives, and priorities for fisheries and ecosystem research, conservation,
and management across multiple fisheries within an ecosystem.
}
