% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{FisheryImpactStatement}
\alias{FisheryImpactStatement}
\title{Fishery Impact Statement (FIS)}
\format{
\describe{
}
}
\source{
NOAA Fisheries; 16 USC 1853(a)(9) (MSA 303(a)(9))
}
\usage{
FisheryImpactStatement
}
\description{
A required provision of a fishery management plan, which assesses, specifies,
and analyzes the likely effects, if any, of the conservation and management
measures on, and possible mitigation measures for: fishery participants,
fishery participants in adjacent areas, and the safety of human life at sea.
}
