% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\docType{data}
\name{Forecast}
\alias{Forecast}
\title{Forecast}
\format{
\describe{
\item{Examples}{short-term forecast}
\item{Rationale}{Future weather is a forecast because it does not require
an if statement, whereas future spawning biomass is a projection because
it depends on an assumption about future catches.}
\item{Synonym or Similar Terms}{projection, prediction}
\item{Range of possible values}{}
\item{Units}{}
}
}
\usage{
Forecast
}
\description{
A probabilistic statement about future events based on what will happen.
}
\keyword{datasets}
