% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{InternationalFisheryAgreement}
\alias{InternationalFisheryAgreement}
\title{International Fishery Agreement (IFA)}
\format{
\describe{
}
}
\source{
16 USC 1802(24) (MSA sec. 2(24))
}
\usage{
InternationalFisheryAgreement
}
\description{
Any bilateral or multilateral treaty, convention, or agreement which relates to
fishing and to which the United States is a party.
}
