% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landings.R
\docType{data}
\name{Landings}
\alias{Landings}
\title{Landings}
\format{
\describe{
\item{Examples}{}
\item{Rationale}{Landings and catch are sometimes thought to be
interchangeable
but they are not given that catch can also include bycatch or
unwanted catch.}
\item{Synonym or Similar Terms}{retained catch, wanted catch}
\item{Range of possible values}{0--Inf}
\item{Units}{mt, numbers}
}
}
\usage{
Landings
}
\description{
Catch that was recorded dockside. Typically, this is catch that was brought
to shore, sorted to species, recorded on a fish ticket after being weighed,
and sold to a processing plant.
}
\keyword{datasets}
