% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumSustainableYield.R
\docType{data}
\name{MaximumSustainableYield}
\alias{MaximumSustainableYield}
\title{Maximum Sustainable Yield}
\format{
\describe{
\item{Examples}{}
\item{Rationale}{Many management frameworks are based on calculations
of MSY despite difficulty in its estimation. Metrics such as fishing rate
at MSY should be abbreviated using all upper case letters but this is not
common practice as of yet. For example, sometimes you will see \code{F_msy} or
some derivative thereof.}
\item{Synonym or Similar Terms}{msy, MSYproxy}
\item{Range of possible values}{0--Inf}
\item{Units}{mt, numbers}
}
}
\source{
insert URL
}
\usage{
MaximumSustainableYield
}
\description{
A maximum level of yield that can be continuously taken from an exploited
stock under average environmental conditions. It can be defined given
biological parameters (e.g., natural mortality and stock productivity) or
using proxies (e.g., the maximum yield at 40\% of unfished biomass). Units
depend on units of catches.
}
\keyword{datasets}
