% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{NationalStandard}
\alias{NationalStandard}
\title{National Standard (NS)}
\format{
\describe{
}
}
\source{
16 USC 1851 (MSA 301); 50 CFR 600.305(a)(3); https://www.fisheries.noaa.gov/national/laws-policies/national-standard-guidelines
}
\usage{
NationalStandard
}
\description{
The Magnuson-Stevens Act requires that a fishery management plan and its
implementing regulations meet ten national standards for fishery conservation
and management. These national standards are principles that must be followed
in any fishery management plan to ensure sustainable and responsible fishery
management.
}
