% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{NortheastFisheriesScienceCenter}
\alias{NortheastFisheriesScienceCenter}
\title{Northeast Fisheries Science Center (NEFSC)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/northeast-fisheries-science-center
}
\usage{
NortheastFisheriesScienceCenter
}
\description{
One of NOAA Fisheries' six Science Centers. The Northeast center provides
science to support the conservation and management of fisheries and living
marine resources across the Northeast Continental Shelf Ecosystem from the Gulf
of Maine to Cape Hatteras.
}
