% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{NorthwestFisheriesScienceCenter}
\alias{NorthwestFisheriesScienceCenter}
\title{Northwest Fisheries Science Center (NWFSC)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/northwest-fisheries-science-center
}
\usage{
NorthwestFisheriesScienceCenter
}
\description{
One of NOAA Fisheries' six Science Centers. The Northwest center provides
science to support the conservation and management of fisheries and living
marine resources across the Pacific Northwest.
}
