% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{PacificIslandsFisheriesScienceCenter}
\alias{PacificIslandsFisheriesScienceCenter}
\title{Pacific Islands Fisheries Science Center (PIFSC)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/pacific-islands-fisheries-science-center
}
\usage{
PacificIslandsFisheriesScienceCenter
}
\description{
One of NOAA Fisheries' six Science Centers. The Pacific Islands center provides
science to support the conservation and management of fisheries and living
marine resources across the Pacific Ocean.
}
