% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{SoutheastFisheriesScienceCenter}
\alias{SoutheastFisheriesScienceCenter}
\title{Southeast Fisheries Science Center (SEFSC)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/southeast-fisheries-science-center
}
\usage{
SoutheastFisheriesScienceCenter
}
\description{
One of NOAA Fisheries' six Science Centers. The Southeast center provides
science to support the conservation and management of fisheries and living
marine resources across the Gulf of America, the South Atlantic, and the
Caribbean.
}
