% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{SpecialManagementZone}
\alias{SpecialManagementZone}
\title{Special Management Zone (SMZ)}
\format{
\describe{
}
}
\source{
50 CFR 622.182(a); 50 CFR 648.148; https://safmc.net/managed-areas/special-management-zones/
}
\usage{
SpecialManagementZone
}
\description{
A designated area to protect fish populations, increase biological production,
and/or create fishing opportunities that would not otherwise exist, through
gear restrictions or other fishing restrictions. SMZs may be designated around
artificial reefs or fish attraction devices, or around spawning sites
(specifically referred to as "Spawning SMZs").
}
