% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{WestCoastRegionalOffice}
\alias{WestCoastRegionalOffice}
\title{West Coast Regional Office (WCRO)}
\format{
\describe{
}
}
\source{
https://www.fisheries.noaa.gov/about/west-coast-region
}
\usage{
WestCoastRegionalOffice
}
\description{
One of NOAA Fisheries' five Regional Offices that supports conservation and
management of fisheries and living marine resources. The West Coast Region
works within the coasts and watersheds of Washington, Oregon, California, and
Idaho. The region covers 317,690 square miles of the eastern Pacific Ocean,
more than 7,000 miles of tidal coastline, and supports ecological functions
within the states’ vast rivers and estuaries.
}
