% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarAcronyms1.R
\name{WesternPacificFisheryManagementCouncil}
\alias{WesternPacificFisheryManagementCouncil}
\title{Western Pacific Fishery Management Council (WPFMC)}
\format{
\describe{
}
}
\source{
Magnuson-Stevens Act, sec. 302 (16 USC 1852)
}
\usage{
WesternPacificFisheryManagementCouncil
}
\description{
One of eight US Regional Fishery Management Councils established by the
Magnuson-Stevens Act. Each Council is responsible for preparing and submitting
to the Secretary of Commerce proposed fishery management plans and amendments
to those plans, as well as proposed regulations that the Council deems
necessary and appropriate to implement the plan, for fisheries within the
Council's specified geographic region that require conservation and management.
The Western Pacific Council consists of the State of Hawaii and the Territories
of American Samoa, Guam, and the Northern Mariana Islands and generally has
authority over the fisheries in the Pacific Ocean seaward of Hawaii and these
Territories.
}
